@@ -24,8 +24,6 @@ classifiers =
     License :: OSI Approved :: MIT License
     Operating System :: OS Independent
     Programming Language :: Python
-    Programming Language :: Python :: 2
-    Programming Language :: Python :: 2.7
     Programming Language :: Python :: 3
     Programming Language :: Python :: 3.6
     Programming Language :: Python :: 3.7
@@ -39,7 +37,7 @@ classifiers =
 packages = find:
 include_package_data = true
 zip_safe = false
-python_requires = >=2.7, !=3.0.*, !=3.1.*, !=3.2.*, !=3.3.*, !=3.4.*, !=3.5.*
+python_requires = >=3.6
 
 install_requires =
     SQLAlchemy>=1.3.0
@@ -91,7 +89,8 @@ ignore =
 exclude = .venv,.git,.tox,dist,doc,*egg,build
 import-order-style = google
 application-import-names = alembic,tests
-
+per-file-ignores =
+                **/__init__.py:F401
 
 [sqla_testing]
 requirement_cls=tests.requirements:DefaultRequirements
