@@ -1,4 +1,5 @@
 from collections import OrderedDict
+from io import StringIO
 import re
 
 from mako.pygen import PythonPrinter
@@ -11,7 +12,6 @@ from ..operations import ops
 from ..util import compat
 from ..util import sqla_compat
 from ..util.compat import string_types
-from ..util.compat import StringIO
 
 
 MAX_PYTHON_ARGS = 255
@@ -480,7 +480,7 @@ def _alter_column(autogen_context, op):
     return text
 
 
-class _f_name(object):
+class _f_name:
     def __init__(self, prefix, name):
         self.prefix = prefix
         self.name = name
@@ -500,17 +500,7 @@ def _ident(name):
     if name is None:
         return name
     elif isinstance(name, sql.elements.quoted_name):
-        if compat.py2k:
-            # the attempt to encode to ascii here isn't super ideal,
-            # however we are trying to cut down on an explosion of
-            # u'' literals only when py2k + SQLA 0.9, in particular
-            # makes unit tests testing code generation very difficult
-            try:
-                return name.encode("ascii")
-            except UnicodeError:
-                return compat.text_type(name)
-        else:
-            return compat.text_type(name)
+        return compat.text_type(name)
     elif isinstance(name, compat.string_types):
         return name
 
