@@ -1,11 +1,11 @@
 import sys
 
-from . import context  # noqa
-from . import op  # noqa
+from . import context
+from . import op
 from .runtime import environment
 from .runtime import migration
 
-__version__ = "1.6.6"
+__version__ = "1.7.0"
 
 sys.modules["alembic.migration"] = migration
 sys.modules["alembic.environment"] = environment
