@@ -1,32 +1,31 @@
-from sqlalchemy.testing import config  # noqa
-from sqlalchemy.testing import emits_warning  # noqa
-from sqlalchemy.testing import engines  # noqa
-from sqlalchemy.testing import exclusions  # noqa
-from sqlalchemy.testing import mock  # noqa
-from sqlalchemy.testing import provide_metadata  # noqa
-from sqlalchemy.testing import uses_deprecated  # noqa
-from sqlalchemy.testing.config import combinations  # noqa
-from sqlalchemy.testing.config import fixture  # noqa
-from sqlalchemy.testing.config import requirements as requires  # noqa
+from sqlalchemy.testing import config
+from sqlalchemy.testing import emits_warning
+from sqlalchemy.testing import engines
+from sqlalchemy.testing import exclusions
+from sqlalchemy.testing import mock
+from sqlalchemy.testing import provide_metadata
+from sqlalchemy.testing import uses_deprecated
+from sqlalchemy.testing.config import combinations
+from sqlalchemy.testing.config import fixture
+from sqlalchemy.testing.config import requirements as requires
 
-from alembic import util  # noqa
-from .assertions import assert_raises  # noqa
-from .assertions import assert_raises_message  # noqa
-from .assertions import emits_python_deprecation_warning  # noqa
-from .assertions import eq_  # noqa
-from .assertions import eq_ignore_whitespace  # noqa
-from .assertions import expect_raises  # noqa
-from .assertions import expect_raises_message  # noqa
-from .assertions import expect_sqlalchemy_deprecated  # noqa
-from .assertions import expect_sqlalchemy_deprecated_20  # noqa
-from .assertions import expect_warnings  # noqa
-from .assertions import is_  # noqa
-from .assertions import is_false  # noqa
-from .assertions import is_not_  # noqa
-from .assertions import is_true  # noqa
-from .assertions import ne_  # noqa
-from .fixtures import TestBase  # noqa
-from .util import resolve_lambda  # noqa
+from .assertions import assert_raises
+from .assertions import assert_raises_message
+from .assertions import emits_python_deprecation_warning
+from .assertions import eq_
+from .assertions import eq_ignore_whitespace
+from .assertions import expect_raises
+from .assertions import expect_raises_message
+from .assertions import expect_sqlalchemy_deprecated
+from .assertions import expect_sqlalchemy_deprecated_20
+from .assertions import expect_warnings
+from .assertions import is_
+from .assertions import is_false
+from .assertions import is_not_
+from .assertions import is_true
+from .assertions import ne_
+from .fixtures import TestBase
+from .util import resolve_lambda
 
 try:
     from sqlalchemy.testing import asyncio
