@@ -1,5 +1,5 @@
 #!coding: utf-8
-
+import importlib.machinery
 import os
 import shutil
 import textwrap
@@ -11,9 +11,6 @@ from . import util as testing_util
 from .. import util
 from ..script import Script
 from ..script import ScriptDirectory
-from ..util.compat import get_current_bytecode_suffixes
-from ..util.compat import has_pep3147
-from ..util.compat import u
 
 
 def _get_staging_directory():
@@ -290,15 +287,13 @@ def make_sourceless(path, style):
 
     py_compile.compile(path)
 
-    if style == "simple" and has_pep3147():
+    if style == "simple":
         pyc_path = util.pyc_file_from_path(path)
-        suffix = get_current_bytecode_suffixes()[0]
+        suffix = importlib.machinery.BYTECODE_SUFFIXES[0]
         filepath, ext = os.path.splitext(path)
         simple_pyc_path = filepath + suffix
         shutil.move(pyc_path, simple_pyc_path)
         pyc_path = simple_pyc_path
-    elif style == "pep3147" and not has_pep3147():
-        raise NotImplementedError()
     else:
         assert style in ("pep3147", "simple")
         pyc_path = util.pyc_file_from_path(path)
@@ -341,11 +336,10 @@ def downgrade():
     write_script(
         script,
         b,
-        u(
-            """# coding: utf-8
+        f"""# coding: utf-8
 "Rev B, méil, %3"
-revision = '{}'
-down_revision = '{}'
+revision = '{b}'
+down_revision = '{a}'
 
 from alembic import op
 
@@ -357,8 +351,7 @@ def upgrade():
 def downgrade():
     op.execute("DROP STEP 2")
 
-"""
-        ).format(b, a),
+""",
         encoding="utf-8",
     )
 
