@@ -1,31 +1,30 @@
-from .compat import raise_  # noqa
-from .editor import open_in_editor  # noqa
+from .editor import open_in_editor
 from .exc import CommandError
-from .langhelpers import _with_legacy_names  # noqa
-from .langhelpers import asbool  # noqa
-from .langhelpers import dedupe_tuple  # noqa
-from .langhelpers import Dispatcher  # noqa
-from .langhelpers import immutabledict  # noqa
-from .langhelpers import memoized_property  # noqa
-from .langhelpers import ModuleClsProxy  # noqa
-from .langhelpers import rev_id  # noqa
-from .langhelpers import to_list  # noqa
-from .langhelpers import to_tuple  # noqa
-from .langhelpers import unique_list  # noqa
-from .messaging import err  # noqa
-from .messaging import format_as_comma  # noqa
-from .messaging import msg  # noqa
-from .messaging import obfuscate_url_pw  # noqa
-from .messaging import status  # noqa
-from .messaging import warn  # noqa
-from .messaging import write_outstream  # noqa
-from .pyfiles import coerce_resource_to_filename  # noqa
-from .pyfiles import load_python_file  # noqa
-from .pyfiles import pyc_file_from_path  # noqa
-from .pyfiles import template_to_file  # noqa
-from .sqla_compat import has_computed  # noqa
-from .sqla_compat import sqla_13  # noqa
-from .sqla_compat import sqla_14  # noqa
+from .langhelpers import _with_legacy_names
+from .langhelpers import asbool
+from .langhelpers import dedupe_tuple
+from .langhelpers import Dispatcher
+from .langhelpers import immutabledict
+from .langhelpers import memoized_property
+from .langhelpers import ModuleClsProxy
+from .langhelpers import rev_id
+from .langhelpers import to_list
+from .langhelpers import to_tuple
+from .langhelpers import unique_list
+from .messaging import err
+from .messaging import format_as_comma
+from .messaging import msg
+from .messaging import obfuscate_url_pw
+from .messaging import status
+from .messaging import warn
+from .messaging import write_outstream
+from .pyfiles import coerce_resource_to_filename
+from .pyfiles import load_python_file
+from .pyfiles import pyc_file_from_path
+from .pyfiles import template_to_file
+from .sqla_compat import has_computed
+from .sqla_compat import sqla_13
+from .sqla_compat import sqla_14
 
 
 if not sqla_13:
