@@ -1,23 +1,21 @@
 import os
 from os.path import join
+from unittest.mock import patch
 
 from alembic import util
 from alembic.testing import combinations
 from alembic.testing import expect_raises_message
-from alembic.testing import mock
 from alembic.testing.fixtures import TestBase
 
 
 class TestHelpers(TestBase):
     def common(self, cb, is_posix=True):
-        with mock.patch(
-            "alembic.util.editor.check_call"
-        ) as check_call, mock.patch(
+        with patch("alembic.util.editor.check_call") as check_call, patch(
             "alembic.util.editor.exists"
-        ) as exists, mock.patch(
+        ) as exists, patch(
             "alembic.util.editor.is_posix",
             new=is_posix,
-        ), mock.patch(
+        ), patch(
             "os.pathsep", new=":" if is_posix else ";"
         ):
             cb(check_call, exists)
